Modularize and Nixify Hyprland Config
First we split up hyprland config to make it easier to manage. Mostly
so I can modify windowrules and keybinds without navigating the
config.

Moved all of the config into a hypr.conf and gave the folder a default
hyprland.conf to source this hypr.conf in non nix systems this will
seem redundant but in nix stow will ignore hyprland.conf because it
already exists