Emacs Upgrades
A lot of these changes are just formatting, but some of them are
noteable:

- Previously dired/dirvish buffers could be added to the mark
ring. This would result in ctrl-o putting me in a file browser rather
than the last file I was in, this was stupid behavior that we now
check for and manually prevent

- killing buffers is now done without confirmation even if changes
 exist. This is something that I prefer to have, if I am choosing to
 kill a buffer I do not want to be asked if I want to do so.

- Lsp keybinds from lsp-mode are now setup to use native eglot
commands and built in xref command replacements with some sanity
checks for when the server eglot is using is incompetent.

- Corfu auto t was removed. Im debating removing popups, trying to
find a balance between my tools helping me and my tools being a
cruch that prohibits and replaces learning

- Formatting before saving is now done with eglot format without
showing any errors if the server cannot do that. and failing that
apheleia should kick in for almost everything else

- jtsx was added, I find it to be a much better typescript / js mode